fix casting to uint in handlers (vacancies)

+++ internal/network/handlers/vacancy.go
@@ -43,7 +43,7 @@ func (vh *VacancyHandler) GetVacancyById(c *gin.Context) {
 		_ = c.Error(errorHandler.ErrInvalidParam)
 		return
 	}
-	vacancy, err := vh.vacancyUseCase.GetById(id)
+	vacancy, err := vh.vacancyUseCase.GetById(uint(id))
 	if err != nil {
 		_ = c.Error(err)
 		return
@@ -58,7 +58,7 @@ func (vh *VacancyHandler) GetUserVacancies(c *gin.Context) {
 		_ = c.Error(errorHandler.ErrInvalidParam)
 		return
 	}
-	vacancies, GetErr := vh.vacancyUseCase.GetByUserId(companyId)
+	vacancies, GetErr := vh.vacancyUseCase.GetByUserId(uint(companyId))
 	if GetErr != nil {
 		_ = c.Error(GetErr)
 		return
@@ -106,7 +106,7 @@ func (vh *VacancyHandler) DeleteVacancy(c *gin.Context) {
 		_ = c.Error(errorHandler.ErrInvalidParam)
 		return
 	}
-	deleteErr := vh.vacancyUseCase.Delete(userId, id)
+	deleteErr := vh.vacancyUseCase.Delete(userId, uint(id))
 	if deleteErr != nil {
 		_ = c.Error(deleteErr)
 		return
@@ -133,7 +133,7 @@ func (vh *VacancyHandler) UpdateVacancy(c *gin.Context) {
 		return
 	}
 
-	updateErr := vh.vacancyUseCase.Update(userId, id, &input)
+	updateErr := vh.vacancyUseCase.Update(userId, uint(id), &input)
 	if updateErr != nil {
 		_ = c.Error(updateErr)
 		return
